Move syntax config into a builder
