+++ examples/generate-yaml/src/main.rs
@@ -2,19 +2,19 @@
 use std::collections::BTreeMap;
 use std::env;
 
-use minijinja::{context, Environment, Syntax};
+use minijinja::syntax::SyntaxConfig;
+use minijinja::{context, Environment};
 
 fn main() {
     let mut env = Environment::new();
-    env.set_syntax(Syntax {
-        block_start: "{%".into(),
-        block_end: "%}".into(),
-        variable_start: "${{".into(),
-        variable_end: "}}".into(),
-        comment_start: "{#".into(),
-        comment_end: "#}".into(),
-    })
-    .unwrap();
+    env.set_syntax(
+        SyntaxConfig::builder()
+            .block_delimiters("{%", "%}")
+            .variable_delimiters("${{", "}}")
+            .block_delimiters("{#", "#}")
+            .build()
+            .unwrap(),
+    );
     env.add_template("template.yml", include_str!("template.yaml"))
         .unwrap();
     let tmpl = env.get_template("template.yml").unwrap();
